+++ example/src/controllers/examples/samples.js
@@ -5,17 +5,17 @@ import mapObj from '../../../../src/helpers/mapObj';
 
 const codeTextFiles = {
     basic: require('fs').readFileSync(__dirname + '/subsections/basic/code.js', 'utf8'),
-    validator: require('fs').readFileSync(__dirname + '/subsections/validator/code.js', 'utf8')
+    signup: require('fs').readFileSync(__dirname + '/subsections/signup/code.js', 'utf8')
 };
 
 const codeComponents = {
     basic: require('./subsections/basic/code'),
-    validator: require('./subsections/validator/code')
+    signup: require('./subsections/signup/code')
 };
 
 export const markdown = {
     basic: require('fs').readFileSync(__dirname + '/subsections/basic/description.md', 'utf8'),
-    validator: require('fs').readFileSync(__dirname + '/subsections/validator/description.md', 'utf8')
+    signup: require('fs').readFileSync(__dirname + '/subsections/signup/description.md', 'utf8')
 };
 
 // Map over each of our codeX imports and make a FormExample for it. The
