+++ example/src/controllers/examples/subsections/signup/index.js
@@ -1,0 +1,4 @@
+export const title = 'Signup';
+export const link = 'Signup';
+export const markdown = require('../../samples').markdown.signup
+export const code = require('../../samples').code.signup
