--- example/src/controllers/examples/subsections/validator/description.md
@@ -1,1 +1,0 @@
-Custom validation on several fields? No problemo.
