--- example/src/controllers/examples/subsections/validator/index.js
@@ -1,4 +1,0 @@
-export const title = 'Custom Validator';
-export const link = 'Custom Validator';
-export const markdown = require('../../samples').markdown.validator
-export const code = require('../../samples').code.validator
